refactor: rename actions for company
